refactor(Layout): Change location of authentication starting line
