+++ src/components/common/Layout/index.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useEffect} from 'react';
 import {Grid, Typography} from '@material-ui/core';
 import {Nav, Container} from '../../common';
 import { useStyles } from './styles';
@@ -6,6 +6,8 @@ import { useLocation } from 'react-router-dom';
 import { matchPath } from 'react-router';
 import {routes} from '../../../utils';
 import theme from '../../../../assets/theme';
+import { useSelector, useDispatch } from 'react-redux';
+import { actions } from '../../../actions';
 
 const Layout = (props) => {
 	const {children} = props;
@@ -13,10 +15,24 @@ const Layout = (props) => {
 	const location = useLocation();
 	const route = routes.find(route => matchPath(location.pathname, route));
 
+	const { userinfo } = useSelector(state => ({ ...state.login }));
+	const dispatch = useDispatch();
+	useEffect(() => {
+		dispatch(actions.getUserInfo());
+	}, []);
+
+	console.log(userinfo);
+
 	return (
 		<Grid
 			container
 		>
+			{userinfo.username !== undefined && <Grid
+				item
+				xs={12}
+			>
+				Logged in as {userinfo.username}
+			</Grid>}
 			<Grid
 				item
 				xs={12}
