+++ src/components/common/Nav/index.js
@@ -5,8 +5,6 @@ import { Link } from "react-router-dom";
 import { useStyles } from './styles';
 import theme from '../../../../assets/theme';
 import {routes} from '../../../utils';
-import { useSelector, useDispatch } from 'react-redux';
-import { actions } from '../../../actions';
 
 const MobileNav = (props) => {
 	const {classes} = props;
@@ -59,26 +57,12 @@ const Nav = (props) => {
 	const classes = useStyles(theme);
 	const isMobile = useMediaQuery(theme.breakpoints.down('xs'));
 
-	const { userinfo } = useSelector(state => ({ ...state.login }));
-	const dispatch = useDispatch();
-	useEffect(() => {
-		dispatch(actions.getUserInfo());
-	}, []);
-
-	console.log(userinfo);
-
 	return (
 		<Grid
 			container
 			justify={isMobile ? "space-between" : "space-around"}
 			alignItems="center"
 		>
-			{userinfo.username !== undefined && <Grid
-				item
-				xs={12}
-			>
-				Logged in as {userinfo.username}
-			</Grid>}
 			{isMobile ? <MobileNav {...{classes}}/> : routes.filter((route) => route.mainLink).map((route, key) => (
 				<Grid item {...{ key }}>
 					<Link {...{ to: route.path, style: { textDecoration: 'none' } }}>
